web: Allow password PDF downloading for 5min
So there is possibility to re download it if the first download fails
for some reason. The timer starts running when the PDF is downloaded for
the first time.

Drop the user_store jobs set as it does not support expirations for
individual values. It was only used for debugging purposes anyway.