rest: Fix (kinda) weird race condition with LDAP::Mod
This is not really a fix. For some unknown reason to me the bug just
does not appear anymore after this change.

It seems that when LDAP::Mod instances are kept long enough in memory in
some cases the keys inside them can change. It's very hard to debug when
it happens because when the values are printed out or a debugger is
started the bug goes away!

For example sambaSID attribute can just change to "k" in the instance
<LDAP::Mod:0x5dfdd10 LDAP_MOD_REPLACE{"k"=>["S-1-5-21-9152723-52889744-89675554-35134"]}>
without any reasonable explanation I can see. The key "k" is a random
ascii char. This then causes "LDAP::ResultError: Undefined attribute
type" errors.

This commit just moves the LDAP::Mod creation closer to the low level
LDAP client method calls and now I'm unable to reproduce the bug.
Awesome.