When deleting a user, do set the user password to some random value.
This seems strange but it does "invalidate" the kerberos principal,
that unfortunately remains even after the user is deleted from ldap.
We should really destroy the kerberos principal, but that is
somewhat more complicated (one problem is that ACLs do not
(AFAICS) permit access for puavo users to the kerberos subtree in
ldap).