First attempt at trying to make external login password changes happen with the permissions of the changing user (instead of the global "root-like" DN in configuration.  The point of this is to make it possible to give only very reduced privileges to the "root-like" dn.
To make this work the change_passwd() interface has been changed to take
only usernames, and not LDAP DNs.  This is because DNs are Puavo-specific,
and we want to do password changes to external databases as well,
possibly even in cases where users do not exist in Puavo at all.

This is not tested much, so fixup commits are likely to come...