Use flexboxes for first- and second-level tab bars.
This fixes numerous problems in the tabs, but it also introduces
one little problem that is fixed with a box-shadow hack: the
bottom border in active tabs should be invisible. But the border
comes from the outer container, not the tabs, so getting a tab
to overwrite the border of its container requires the box-shadow
hack. But I guess the other things flexboxes fix will be worth it.