Fix broken user tests.
I'm not sure when I did it, but I've broken some CSS selector
somewhere, because these tests will fail with the error "Ambiguous
match, found 2 elements matching visible link...". Debugging
the feature with byebug and pry did not reveal anything out of
the ordinary, so the true cause of this remains a mystery.