Show the password requirements on username list pages.
The lists themselves currently lack all validation. It is possible
to even try to set empty password. But adding validation to these
forms is very difficult and time-consuming. Should not be a very
big problem, because very few people use this page, and we can
assume they know what they're doing.