Revise the integration "badge" system.
The original intention was add support for displaying the next
synchronisation timestamps in the UI and it ballooned from there.
It still won't display the next time, but that comes later.

Now the integration display is completely optional. If integrations.yml
does not exist or it cannot be loaded, nothing bad happens.