Don't put everything from the User object in the SSO URLs.
We've had lots of problems with SSO because the URLs grow too long
when users are added to "too many" groups. The solution is filtering.
Don't just cram the whole User object in there, instead only put in
the parts that are actually part of our SSO spec. Plus a few other
extra fields that we don't know if anyone is using.

This can potentially break everything.