Remove some obsolete JavaScript from application.js.
Some of this stuff hasn't worked in months now, but no one has
reported of anything being broken, so I guess it's safe to
remove them.

The "advanced printer permissions" stuff goes back to 2013 and
it hasn't been touched since. I can't even determine what it
was. But what I *can* determine is that putting it's JS on a
file that is included on every page is wrong. That's not how
it's done. You put per-page JavaScript on the page where it is
needed.