Version 0.1 alpha of the long-waited dynamic table sorting system.
This JavaScript thing turns pretty much any table into neatly
sortable table. You can click the table headers (it supports
multiple header rows) to sort the table by that column in either
ascending or descending order.

This still needs a *lot* of work. For example, it breaks mobile
layouts if even one of the table headers are clicked.

There are libraries that do this, but none of them work directly
with our tables, so I hacked together our own solution.