Downgrade Sinatra to 2.0.4
There's a bug in Sinatra 2.0.5 that makes one REST test fail
(devices_write_test.rb) as it tries to dup a NilClass. See

https://github.com/rack-test/rack-test/issues/243

for discussion.

The commit that fixes the error is

https://github.com/sinatra/sinatra/commit/d8c18394d86626726c256ce57c3336ca649a33ae

but, at the time of this commit, there have been no Sinatra releases
in almost nine months and the fix is not in any released version.
So downgrading is the only option.

EDIT: Just after making this commit, I noticed they released
a fixed version yesterday (2.0.6) and another fix (2.0.7) this
morning. I am not upgrading it yet.