Remove support for multiple search URLs in LiveQuery search.
They caused mysterious problems in Rails 5, making the top bar search
box simply not work. The search box on group edit pages worked fine.
Since multiple URLs are not needed (more commits will follow) anymore,
the support for them can be removed.

This JavaScript file is orphaned anyway. It's original author no longer
maintains it, nor lists it on his GitHub page. So it's safe to assume
we can modify it as much as we want.