Hide the "change someone else's password" tab if wanted.
The "&hidetabs" is set in Puavomenu. Non-teachers/non-admins can't
change other people's passwords, and since most people access the
password form through the menu, it is therefore a good idea to hide
the other tab by default. If accessed through
<organisation>.opinsys.fi, then the other tab is displayed normally,
because most people who use that address will be teachers and
admins.