Look up password requirements in functions that need them.
There was a bug on user edit forms which hid the password requirement
hint texts and password validators if the form did not validate.
This happened because we didn't look up the requirements again. So do
that in the two functions that actually need the information rather
than passing it to them as arguments. No more missing things.