AJAX endpoints for getting users, groups and devices dynamically.
One of the many upcoming commits that are part of the new filterable
and dynamic user/group/device tables.

It will still take a long time until they fully replace the current
indexes, but I need to start from something. The first parts of this
functionality were written over a month ago.