The Supertable (TM)
This JavaScript -based table thingy can dynamically reload itself,
select which columns are shown (and even in which order!), filter
rows on one or more filter criterias, export its contents as CSV
and more.

The TODO list for it is still long, but this will get it started
nicely.

It is the largest piece of JavaScript I have ever written.

It is also a monster.