Convert school devices to use the supertable
There are now two index files: one for the supertable, one for
the tests. Despite of several attempts, I have not been able to
enable JavaScript in Cucumber/webrat/spec tests, so many tests
will fail if the page contents depends on JavaScript. So this
ugly hack is required, for now. The same problem also affects
groups and users.