Use Ruby's own JSON parser here
I tracked down the original commits from December 2013 that
introduced yajl to puavo, but they were related to things
that either aren't used anymore, or are probably fixed by
now. So yajl can go. It stopped working anyway when fluentd
was removed, as the require's came through the fluent wrapper.