Synchronous user deletions from Google
Strictly speaking, this code does not belong in a public repository,
but we already have Google-related code available publicly.

And you still need a server that'll respond to the messages sent
by this function and *that* part is not public. If you really need
this functionality, take a look at Google's Admin SDK and build
your own server. It can be done in <100 lines of Python/Ruby.