Initial commit for the experimental V4 REST API helper functions.
CAUTION: This will burn you. Do not use it unless you're prepared
to treat severe burn wounds.

This is alpha-level stuff. Perhaps one day it will be a proper,
actually usable API. But for now, it is primarily meant for internal
testing and experimenting with low-level LDAP filters and searching.

The idea for this came from our third-party integrations. They
retrieve all users in the organisation, in bulk, diff them with
other data, then save the changes. And that bulk fetch is the
slowest part. One of our customers, for example, has almost 9000
users. The old "v3/users" call can churn that for some 6-7 minutes
and consumes a gigabyte of memory (yay, Ruby) before returning
anything. This new API does the same thing in ~2 seconds and use
only about 30-40 megabytes of memory. Of course it won't return all
of the same information, but it returns the information *we actually
need*.

This code is NOT FINAL. It WILL CHANGE without any prior warnings!
It can be deleted without a warning!

It is also read-only at the moment. Creating and updating things
will be implemented later (some of it is actually done already,
but not needed yet).