We only care about the member count, not the members themselves
Accessing school.members does a VERY slow LDAP search for all the
members, while school.memberUid just searches for the usernames
and that's all we need for the user count.

Except... in some schools these two members are not entirely in
sync. You can count the differences with your fingers usually,
so it's not terribly fatal.