Do not call User.to_hash when generating sessions
It wastes memory and CPU time and poses a (small) privacy problem,
because your profile JSON will, for example, list all the groups and
schools you're in, with their username lists. You don't need that
information.

This commit has a potential to wreak havoc on epic scale. The data
returned from a "POST /v3/sessions" call is used in, among other
places, when logging in (by the puavo-login script). If the data
isn't correct, you won't log in to your desktop. I've verified every
piece of data, but it's possible (read: very likely) that something
has slipped through and is now missing.

We'll see how long it takes for this commit to be reverted.