Speed up device external files listing and retrieval
These endpoints are called *a lot* (almost 100k times in two weeks),
so I surmised that if they are optimized to not construct any Device
or ExternalFile objects, then they should execute faster and take
less memory, right?

This is a commit that can potentially break everything. I spent
over two hours testing this. It returns the same data as the old
calls.

If this breaks something, or nothing happens to memory use, I will
revert this commit.