Fix some ACL tests by cleaning up ldap when establishing a new test environment.
The "separate" environments actually have a common environment
in ldap level, thus if some user/device gets created twice,
this will not work without clean in between.

Also provide a way to access the model object through env.
Accessing "school" when creating "minerva.mcgonagall" needed this,
using @school did not work because that would not be specific
to the environment (cleaning up ldap would mean that @school
would be a stale reference).