Gem update
First proper gem update in almost a year. There were smaller
updates, but they were limited because we were stuck on Stretch.
Now this is the first full Buster-era gem update and that shows.
All tests pass, been in use for two days now. Further commits
related to this will follow, so if this has to be reverted,
take care of the other commits too.