Enable secure SameSite cookies in development and production
Every single test will fail with this because the test system
cannot login in. I don't know why, but it only affects tests
so it's easy to fix.

Whether this change actually fixes the core issue, I don't know.
But the warnings in browsers go away and logins still work,
so maybe it's good?