Remove the "msgtag" parameter from all logger calls.
We haven't used Fluent in a few years now and we do not have any
ElasticSearch instances where these "tags" could be searched in
anyway. So remove them.

I hope I did not introduce any bugs when I rewrote some of the
log calls. Some of them are almost impossible to test, so if
there are typos/bugs in them, they will explode in production.