Redesign how the device primary user information is handled and saved
If you opened puavo-web-prompt, searched for a device, then saved it
without changing anything, the device's primary user, if any, was
cleared. This is because the model code expected the "primary_user_uid"
member to be set. And if you didn't the code happily removed the current
primary user. This ugly code fixes that.

It took almost 13 hours in total to locate, debug and fix this bug.
I hope it works.