Revert "Restart puavo-rest before running rest tests."
This reverts commit 1b7fd5a53e457318e49820a66a5effb9dd4b40b8.

This did not actually help.  One tests fails for some reason,
and restarting puavo-rest helps with that.  And I actually tested
this change as is before committing, it appeared to solve the
problem.  But now the problem appeared again even with this change.
So reverting, this does not look sensible anyway.