Add a custom 404 page in production mode
I have no idea if this is the right way. There are multiple
ways to get 404 pages in RoR. And not a single one of them
worked. The "catch all" route was the last measure.

There is a 404.html in the public directory. I don't know
why it does not appear by default, nor how to make it appear.