puavo-rest: Remove duplicate pending mods before save
This is a strange bug. If you create, for example, a new user
and specify one or more of its attributes more than once, like
this:

    u = PuavoRest::User.new
    u.first_name = "Test"
    u.last_name = "User"
    u.username = "test.use"     # oops, a typo
    u.username = "test.user"    # there, fixed
    u.roles = ["testuser"]
    u.school_dns = ["..."]
    u.save!

Then the save will fail with a message "Type or value exists"
and it means there is now more than one pending modifications
targeting the "username" attribute and for some reason, the LDAP
library really REALLY does not like that. Deduplicating the
@pending_mods array fixes this.

This only seems to affect new entries. I can't reproduce the
problem when editing existing entries. But that does not mean
really anything, maybe I just tested it wrong?

All tests pass with this change, but given how deep it is in
the bowels of the system, I am legit afraid to make this commit.
But it is needed.