Add an automatic/enforced email address generation mode
Some organisations want users to have non-editable, automatically-
generated email addresses with custom domains, so add support
for that.

Retrofitting a system like this to an old codebase is not my
area of expertise, but it seems to work. I am slightly concerned
about how much this slows down certain operations, but I can't
really do anything about that now.