When getting all organisation users, use raw queries for schools
This offers significant speedups, as School.all can be very slow
in some places. The database query itself is still the slowest
on average, but in one organisation, I measured School.all taking
longer than the database query for 11000 users! That's because
every School object contains also a list of their users as objects,
so we get multiple layers of searches here. None of that is needed,
so use raw objects and get only the pieces we actually need.

There are more optimization possibilities in the data conversion
and JSON conversion (yes, it is slow), but those will be done
later.