Use raw queries when creating the "my school users" page
Raw queries require lots of messy code, but the speed improvement
is worth it. Instead of waiting 30 seconds for the page to open,
you only have to wait 2 seconds. Yes, I actually measured the
speed in two largest schools I could find in our database. Both
schools had thousands of students.

I've wanted to do this for two years now, but ACLs prevented me.
But now the new ACLs are in use in all organisations.