Remove the puavoRule field and revamp printing permissions lists
The puavoRule field is a remnant of some very old code (dating
back to 2010-2013) that was never really used for anything, so
remove it.

Since there's nothing left to edit, this commit also changes the
"edit" page to an overview page that lists the special permissions
for the printer. This list will eventually be merged into the
main printers list, but this will do for now.