Send all user, group, device and server columns
This effectively undoes one of the big changes introduced in SuperTable 2
(sending only the data that is visible). The reasons for this undoing are:

1) It didn't speed up the process as much as I thought it would. It did
   speed it up, but not enough to warrant lots of extra code. Testing in
   some of the largest organisations we have showed that sending all
   columns took about 1.5-2.0x the time it took to send only the visible
   columns. But that was the *worst* case. Most organisations are smaller.

2) It made the code messier than it should be.

3) Some people have filters that targe columns that aren't visible and
   this broke those filters (because filters targeting invisible columns
   were skipped).

4) The upcoming filtering system updates to SuperTable 2 requires all
   columns to be sent anyway (to be useful), so I'm preparing for the
   future.

5) The first version of SuperTable sent everything and it worked in
   production just fine for 1.5 years.

6) If it's too slow, throw more server hardware at it.

Sending only the visible columns did reduce the amount of data sent
over the network, but that was about the only real advantage it had.

Ruby and JavaScript aren't the fastest languages around and most of
the speedups are lost in their inefficiencies.