Admin restrictions should only apply if there are blocking integrations
After some brainstorming, we decided to make the "create new users" and
"delete existing users" admin restrictions/permissions to only apply to
schools that have "blocking" integrations enabled. Certain integrations
are sensitive to user creation and deletion and we want to restrict
those schools, but in other schools admins are free.

Eventually we will fully implement proper access rights, but this is
not yet that.