Add support for full row (middle) click opens
Now it's enough to just middle-click on the row to open it for viewing.
Middle clicking the checkbox (if it exists) does nothing. You have to
provide a callback function that formats the URL to be opened, or NULL
if it can't be done.

Due to browser security settings, where the item actually opens in is
uncontrollable. It could be in a new tab, or a new window, or even in
the current tab. There's nothing the script can do about it. In Chrom*
browsers they usually seem to open in background tabs. In Firefox it
opens a new tab and immediately switches to it unless you set
"browser.tabs.loadDivertedInBackground" to "true" in about:config.