Add a simple filtering engine to V4 API calls
Allows "starts with", "ends with", "contains" and "is" string
matching. Multiple filters can be specified with the array syntax
("filter[]=...&filter[]=....&..."), but how well they work in
practice is unknown.

Each filter is made up of three parts: target, operator and value.
Targets are field names, like "username", "abbreviation" and so on.
Operator is "starts", "ends", "contains" or "is". And the value
is what you're searching for. They're separated with pipes, because
pipes are unlikely to appear in legit search values. If you need
to search for a pipe, you're out of luck. Do coarse filtering when
requesting data, and then finer searching in your client program
where the URL in-band signaling restrictions don't apply.

For example, to search for users whose name starts with "foo":

/v4/users?fields=...&filter=name|starts|foo

To search for teaching groups whose name contains "test":

/v4/groups?fields=....&filter[]=type|is|teaching%20group&
  filter[]=name|contains|test

And so on. Use HTTP escapes (like %20 for space) when needed.