Clean up and rename LdapModel.escape
The method isn't actually called (the overloaded Rack utility method
escape() was called instead and it did a HTML escape, not LDAP escape).

I don't even understand why the method existed in the first place.
It was a verbatim copy-paste of Net::LDAP::Filter.escape. Perhaps
there originally was a reason for it, but it was never documented
anywhere? All the commits relating to this date back to 2013.