Don't crash if the current user can't read the puavoSchool attribute
Here's the thing, non-owners cannot access the school index page, so
those users who can access it, can also see the puavoSchool attribute.
Except... the check if the current user should see the school page
isn't done until AFTER the data has been collected, in the "respond_to"
block. So they end up accessing what they should not access, but the
results are then discarded.

This needs to be fixed properly, but I can't be bothered with it right
now.