New rule-based password validation system
No more hard-coded validation logic, all rules are specified in a
YAML file on the server. Currently passwords can be verified using
length (min, max, etc.) and regexp-based checks. All rules must
pass when a password is verified, otherwise it is rejected. It's
possible for a password to fail multiple rules, in which case you
will get multiple validation error messages. This is by design.

The rule verifier has no mercy: if you write incorrect rules the
server won't boot. If you refer to a non-existent ruleset in an
organisation's integrations, the server exits. So don't write
invalid rules or use invalid ruleset names. (The primary reason
for this is that it's very difficult to properly deal with errors
in a Rails' initializer. Exiting the process is the only sane
choice.)