Determine and propagate the primary school ID on password change forms
Password validation is difficult because the form does not know
to which school the target user belongs to, and thus the password
complexity cannot be checked. But once the user data is retrieved
from the database, then we do know it. Stash it in a hidden field
on the form. If the school ID is known when the form is initially
opened, then all the better, we can display the requirements
immediately.