Reject common passwords
There's a short built-in list of common passwords, but if a
larger list file (one word per row) exists, it will be used.

I know this is a bit controversial, but this feature is enabled
everywhere by default. There's currently no way to turn it off.

I left puavo-rest out of this, because user creation and password
changing through puavo-rest is somewhat special and limited.