Store full user info in Redis and use hashed JWT as the key
The JWT cannot be decoded on the puavo-web server because it does not
know the secret. But we can use its SHA384 as a key to Redis and put
full user information there. If the URL is edited, the reset form
won't open as there's no data in Redis matching the edited URL and
the Redis cache cannot be manipulated from the outside.

Now we can customize the password form and validate the password
properly because we know the school the user is in. And because
the Redis entry is removed when the password reset is done, the
link can be expired immediately, no need to wait 10 minutes.

The downside is increased complexity. Fortunately there's not much
left to do here.