Add session support to SSO
The generated session is stored in Redis, and the key is a 128-character
random hexadecimal string, which is stored in a cookie. If the browser
sends this cookie to the login form, the session data is loaded from Redis
and, if it exists and is valid, the user is immediately "authenticated"
and forwarded back to the service that requested the login.

This session mechanism is not enabled by default. It has to be enabled
per-organisation and per-service, when needed.

If a user is edited or removed in puavo-web, the stored session, if it
exists, is immediately invalidated.

The session lasts 8 hours.