Embed group selection directly on the user editing form
No more separate "year class" / "administrative groups" page
that appears after creating a new user. Now all groups in each
school the user can access are listed on the page, and the admin
can simply check the groups the user should be a member of.

The new code no longer uses puavo-rest proxy to set the groups.
It also can handle cases where the user is a member of groups
in schools the admin cannot access. Previously, hiding the
inaccessible schools was a dataloss bug.