User mass import and update tool III
Here it is, finally. The third iteration of the mass user creation
and updating tool.

It's still unfinished. It needs *a lot* of polishing and testing.
But it should do its job just fine.

Nah, I did not split this into dozen smaller commits. Why would I?

The "old" new import tool will be removed once this has been debugged
and finished.