rest: Preserve user groups when the username changes
This is very hacky, but it works. But I cannot guarantee that
it works in all situations. I think it's fairly easy to create
a situation where it tries to add the user to missing groups
or to groups where they already are a member of, despite the
name change.

All existing tests pass. But there are no specific tests for this.