Add a button for forcibly resetting a user's SSO session
This is one of those things someone might use once a year. But if
you know what this is for, it'll save your day when you need it.

Basically, if SSO sessions are enabled in the target organisation
and the target user has an active SSO session, this button will
delete that session, forcing them to re-authenticate. If the user
does not have an SSO session, then nothing will happen.