Allow username lists to be used as a source data for mass updates
This is the reverse of creating a username list. If you didn't
specify passwords or groups (or something else) for newly-imported
users, but you did create a username list, no problems. Just load
it back and specify the columns you left out, then update the
users. The old import tool could not do this trick.