Use the super account proxy for (most) import data retrieval operations
Now the access level of the user does not matter. If the super account
is correctly configured, all import operations will still work fine.
The downsite is massively increased code complexity (the complexity of
new_import.js is starting to reach monstrous levels).

This commit moves most data processing to server-side. This way, we
don't have to send megabytes of usernames back to the browser; we'll
only send the results back. This will also prevent data leakages.
Except the "compare non-username attributes" command will still send
data back to the client. But I don't know yet how to fix that.
I either have to remove that button, or rearchitect it thoroughly.

We can *almost* enable the import tool for all users now. Almost.