Create the username list in the correct school
Don't use the logged-in user's primary school, but the school
where the import tool was opened in.

It's strange, there are other tools that create username lists
and they always go in the right school. Why don't these? What's
the difference? The fix is only two lines, but they shouldn't
be needed at all.