Hack to make the parse groups tool work nicely
If you click the icon, it opens the normal set group dialog,
because "e.target.id" points to the icon element, not the A
element, and thus there is no ID.

The icons should be amended with "pointer-events: none" but
I'm a bit afraid to do that now, because it's possible
something else breaks. It needs to be tested.