Enable the new mass import tool for non-owners
We don't blindly enable it for *all* admins, but only to those we want.
This unfortunately required hacking of the permissions system, as it
wasn't meant for this (we REALLY need that new permissions system and
we need it now).

The tool works when used by a non-owner. Some of the error messages
they can get can be cryptic, since they cannot access/update some
attributes in certain scenarios. But most of the tools' functionality
should work just fine, provided they only try to access their own
school's users. Multi-school experience can be tricker and more
varied, but it should work (I did test it). There are almost certainly
bugs in this related to multi-school admins. We'll fix them later.

Better say goodbye to the old tool. It won't be around much longer.