Make the external services pages read-only and limit their visibility
From now on, only activated services are listed, and the page is
completely read-only. Not even organisation-level owners should be
able to edit them; the potential for information leak is too high.

To edit them, you have to actually use puavo-web-prompt. A new
management script will be built for them soon.