New script for managing external services
First early prototype. Can be used to list services, and
activate/deactivate them, one by one, on organisation and
school level.

Some future version might add such exotic features as
creation of new services (ie. subsumes the functionality of
puavo-add-external-service.rb), deletion of old services,
and more.

The biggest challenge faced when I wrote this (yesterday)
was authentication. When modifying organisation and school
services, you obviously need to connect to that organistion.
But when listing services, no organisation is needed. That
is why there are two different authentication methods. I
want to clean up that mess in the future.

I'm also not 100% sure if --ldap-master and --ldap-user
switches work as advertised.