Move i18n.js to assets
This file has been in public/javascripts/ for years, but due to
recent gem changes, it must now be moved to the assets directory.
I'm not even sure where this file is coming from, it has no
git history. I wonder if there's an updated version of this file
somewhere? I guess it was originally there for the previous
import tool, but it has since been used in other places too,
so it cannot be removed.