Update the Sprockets gem
This was a complex operation and I've been putting it off for
a few years now.

I have *no* idea if I'm doing these changes correctly. Everything
appears to work. Tests pass. I can edit CSS and JavaScript file and
the changes apply. "make deb" works (but I don't know if the
resulting .deb file works).

I'm scared. This is one of those things that can (and will) break
everything if it doesn't work, and there will be more commits like
this in the near future.