Fix the broken search icon on my school users page
I don't know how long this has been broken. I don't know if the
Sprockets gem update somehow broke this. But it was broken.

The original URL works in the development environment, and the
new URL works in production. I have no idea what causes this
difference, because all other resources work in both environments.
I'll prioritize the production environment.

(This isn't a big thing, but it was/is broken.)