SuperTable v3.0
And here is the main meal of the night. This is what kicked off
the whole modularisation process over two weeks ago, and it won't
end here.

SuperTable 3.0 is the same as the last version of SuperTable 2,
but large parts of it has been modularised.

It also brings in significant UI upgrade, as it has been reduced
to a narrow strip above the column names, and you can expand
and pop-out things from it when you need them, they're not always
visible. And they scroll with the page (at least in Firefox,
there's a bug in Chrom* that prevents it from working, but I have
some upcoming hacks that can fix it, *deep sigh*), so you never
lose them. You can now see more on the page, as 1/3 of it isn't
taken up by table controls.

The version number bump is justified, because now that it's not
in one large monolithic file, it can actually be refactored and
worked with again. I have numerous plans for it for the future.
Just you wait...