Don't require logging in to verify an email address
This is horrible. I didn't want to do this. But there's something
weird in Rails that makes everything fail. Usually, Rails remembers
the URL you were trying to open when it presents the login screen.
But not in this case. The verification URL is forgotten and after
you login, all sorts of weird things happen. You *can't* verify
email addesses if you have to log in. So the system has been changed
so that the verification form can be used without logging in.
Now anyone can verify any address, as long as they can get their
hands on the unique token.

I really hope this works... (I also hope I could revert this.)