Remove the home directory attribute from the user interface
It cannot be removed completely, as the posixAccount class (from which
the user class is inherited from) is a standardized LDAP class which
*requires* the attribute to be present and updated. We have to give
it a value, but we won't display it anywhere anymore. Some REST
responses can return it, but eh.

I didn't try what happens if you set it to nil or empty string. I don't
dare touch the standardized LDAP attributes, as I don't fully understand
how they work.