Repurpose the 'puavoServiceTrusted' attribute for more secure SSO
If you ask me, the old puavoServiceTrusted flag was a mistake. There should
be NO hidden services that are implicitly activated everywhere.

This commit removes support for the old functionality (and fixes the tests)
and repurposes the flag for "verified SSO" system. The verified SSO is
otherwise like the normal SSO, but it rejects logins from users who do not
have at least one verified email address.

It took me ages to fix the tests. They were written with the assumption that
the test service was "trusted", ie. always active. Since this isn't true
anymore, the service/services have to be explicitly activated in all test
organisations.