SuperTable: Remove TableFlag constants
It's ugly and not really "in the spirit" of JavaScript.
There's also another thing: we specify the same settings
almost every time a table is created. Why not make them
defaults and then specify the exceptions? This is exactly
what this commit does.

This commit also fixes an annoying bug in the table control
row: tool and filtering sections are removed if they're
completely empty.