SuperTable: Don't duplicate the table data when filtering and sorting
Instead of making a copy of "this.data.transformed", use a lookup table
that iterates over the visible rows in the order they should be. This
offers major memory usage reductions. The downside is that now every
table lookup is indirect. Not that we're trying to optimize for CPU
cache usage here...

This commit also contains a bugfix. The sort stabilization was meant
to compare the PuavoIDs, but instead it accidentally subtracted an
array from another array and coerced the result into... something?
JavaScript was like "yeah, we cool". But not anymore.