Add puavoNotes support to organisations, schools, groups, users and devices
The field is meant for completely free-form notes about the entry. It's not
description. It's *notes*. Two different things.

This commit also adds the school description attribute to the user interface.
For years, it has only existed on the edit form and nowhere else. Now it is
actually displayed and can be used.

This includes even tests for puavo-web and puavo-rest. Those took longer
than everything else! No tests for adding/editing organisation notes, as
I couldn't find a way to do that in the test environment.

I originally intended to add this in multiple commits (one for organisations,
one for schools, one for users, etc.) but then I realized it's pointless,
so here's the whole thing in one commit.