Rails 6.1 no longer renders templates with absolute/relative paths
Use symlinks as a workaround. Let's hope they work in production.

This is a stupid change in Rails, if you ask me.

Now it can be argued that these paths should always have been
symlinks, but the point is, removing support for absolute and
relative paths in "render" accomplishes *nothing*. If a file
has been placed outside of the Rails' view lookup magic, then
that was done *on purpose*. The framework should just render
the file it has been told to render and not be pointlessly
obturant.