Fix VirtualHostBase for newer Rack versions
This is only used in one place. Let's hope this really does work in
production. All tests pass with this change.

Basically, the new Rack versions change how the authority is handled,
and if we don't set the HTTP host header correctly, the port number
will disappear, leading to incomplete/broken URLs.

This required almost two days of research.