SuperTable: Don't crash if the table cannot be sorted
It wouldn't be so bad, but if this function crashes, the table will
be completely frozen. It cannot be sorted again. Even reloading
the page won't help, because the sorted column has already been saved
in the settings and they get reloaded when the page loads, so you
can't even change the sort column. Unless you know how to clear the
localStore data for the page, you'll be stuck.