Fix the forced SSO session reset button in puavo-web
In puavo-standalone, puavo-web and puavo-rest use different Redis database
numbers, so in order to test this, you have to temporarily change the Redis
database number in puavo-rest to 1, then perform a login while sessions are
enabled. Then the logout will work. This isn't a problem in production, as
both puavo-rest and puavo-web use the same Redis database.