Gem update (cucumber and rack)
Rails needs webrick but it has no dependency on it. Previously
webrick was installed because cucumber depended on it, but
the new version does not. So I have to manually add it to
dependencies. I don't know.

rack 3.x refuses to install and I wasn't able to determine why.
It's incompatible with... something.