Start supporting Argon2 in external service authentication
Since all Argon2 hashes start with "$argon2", we can distinguish
between hashed and non-hashed secrets. Eventually we won't support
plaintext secrets at all anymore, but we're not ready for that yet.

puavo-add-external-service has no support for this yet, so the
secret has to be hashed in advance, or changed in puavo-rest-prompt.
That will be fixed eventually.