Remove the remaining Eltern code
All the production Eltern services were shut off a long time ago.

(Though, there is one little detail in Eltern I want to resurrect
one day: the domain "overlay" in the username box. The way it was
done here worked but it was a bit hacky. I want to do it properly
one day and put in the normal login forms.)